requirements.txt
requests==2.31.0
scikit-learn==1.3.0
pandas==2.0.3
google-cloud-vision==3.4.4